backend /requirements.txt
Django<4.0,>=3.2
djangorestframework<4.0,>=3.12
psycopg2-binary<3.0,>=2.9
requests>=2.26,<3.0
gunicorn>=20.1.0
